% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunck.R
\name{chunck}
\alias{chunck}
\alias{partition_ok}
\title{Partition data across a cluster}
\usage{
chunck(.data, ..., cluster = NULL, ok = FALSE)

partition_ok(.data, ..., cluster = NULL, ok = FALSE)
}
\arguments{
\item{.data}{Dataset to partition}

\item{...}{Variables to partition by.
Will generally work best when you have many more groups than nodes.
If omitted, will randomly parition rows across nodes.}

\item{cluster}{Cluster to use.
The default is to call \code{\link[multidplyr]{get_default_cluster}()}.}

\item{ok}{logical.
If \code{FALSE}, no partition is applied,
\code{.data} is returned unchanged.}
}
\value{
The dataset partitioned accross a cluster.
}
\description{
The function \code{chunck} is identical to
\code{\link[multidplyr]{partition}} from package \pkg{multidplyr},
except that it adds a logical argument \code{ok};
if \code{ok=FALSE}, no partition is applied,
the input dataset is returned unchanged.
}
\author{
Straightforward adaptation of Hadley Wickham's 
\code{\link[multidplyr]{partition}}.
}
\seealso{
\code{\link[multidplyr]{partition}} from package \pkg{multidplyr}.
}

