% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports_bazar.R, R/reexports_dplyr.R,
%   R/reexports_magrittr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{concat}
\alias{reexports}
\alias{concat0}
\alias{reexports}
\alias{concat_}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{mutate_}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{rename_}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{select_}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{transmute_}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bazar}{\code{\link[bazar]{concat}}, \code{\link[bazar]{concat0}}, \code{\link[bazar]{concat_}}}

  \item{dplyr}{\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{rename_}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{select_}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{transmute_}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

